Implemented a majority of the offsetFromGm AMTLV processing

+++ src/main/java/edu/unh/artt/core/error_sample/representation/TimeErrorSample.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface TimeErrorSample {
     double [] getSample();
-    int getWeight();
+    long getWeight();
     String getIdentifier();
     int getNumDimensions();
 }
